backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
PyJWT==2.9.0
python-multipart==0.0.12
scikit-learn==1.5.2
numpy==1.26.4
joblib==1.4.2
